+++ js/random-functions.js
@@ -1,0 +1,22 @@
+
+const getRandomInteger = (min, max) => { //функция получения случайного числа в диапазоне
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+function createRandomNorepeat (min, max) { //функция получения уникального случайного числа в диапазоне
+  const previousValues = [];
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+export {getRandomInteger};
+export {createRandomNorepeat};
